Move elastic logic to specific domains
