+++ src/Database/Model.php
@@ -2,15 +2,16 @@
 
 namespace Baka\Database;
 
+use Baka\Contracts\Database\ModelInterface;
 use Baka\Database\Exception\ModelNotFoundException;
 use Baka\Database\Exception\ModelNotProcessedException;
 use function Baka\getShortClassName;
 use Phalcon\Mvc\Model as PhalconModel;
 use Phalcon\Mvc\Model\ResultsetInterface;
-use Phalcon\Mvc\ModelInterface;
+use Phalcon\Mvc\ModelInterface as PhalconModelInterface;
 use RuntimeException;
 
-class Model extends PhalconModel implements ModelInterface
+class Model extends PhalconModel implements  ModelInterface
 {
     /**
      * Define a model alias to throw exception msg to the end user.
