+++ src/Elasticsearch/IndexBuilder.php
@@ -2,13 +2,13 @@
 
 namespace Baka\Elasticsearch;
 
+use Baka\Contracts\Database\ModelInterface;
 use Baka\Database\CustomFields\CustomFields;
 use Baka\Elasticsearch\Model as ModelCustomFields;
 use Exception;
 use Phalcon\Db\Column;
 use Phalcon\Di;
 use Phalcon\Mvc\Model;
-use Phalcon\Mvc\ModelInterface;
 use ReflectionClass;
 
 class IndexBuilder
